requirements.txt
Django==1.9.4
pinax-theme-bootstrap==7.4.0
pinax-types==0.10.0
python-dateutil==2.5.3
